Move GameInterface to its own file
